refactor: split single file into modular architecture
Split the monolithic main.ts (213 lines) into 4 focused modules:
- types.ts: Type definitions (Options, RemoteRepositoryServiceUrls)
- utils.ts: Utility functions (getPattyRoot, getPattyDirs, repository operations)
- commands.ts: Command implementations (create, get, list, root)
- main.ts: CLI definition and entry point (37 lines)

This improves code organization, maintainability, and separation of concerns while maintaining all existing functionality.
